exercicio05

+++ src/main/java/com/mycompany/poo/aula11/Pessoa.java
@@ -36,24 +36,20 @@ public class Pessoa {
     }
 
     
-    public int getSalario() {
+    public double getSalario() {
         return salario;
     }
 
-    public void setSalario(int salario) {
+    public void setSalario(double salario) {
         this.salario = salario;
     }
  
-    public int getAltura() {
+    public double getAltura() {
         return altura;
     }
 
-    public void setAltura(int altura) {
+    public void setAltura(double altura) {
         this.altura = altura;
     }
-
-
-
-    
 
 }
